refactor(telegram): split handlers.ts into handler modules
handlers.ts was 1439 lines mixing access control, message queueing,
streaming delivery, and per-type message handlers, all sharing
module-level state. Split into a handlers/ folder with a canonical
layout (shared with the discord frontend):
  - state    — shared maps/sets (access config, queues, rate windows)
  - access   — DM whitelist + group admin checks, unauthorized handling
  - context  — sender/reply/forward context + Telegram file downloads
  - delivery — HTML send, streaming drafts, processAndReply pipeline
  - queue    — per-chat debounce queue + per-user rate limiting
  - messages — per-message-type handlers

index.ts re-exports the identical public surface; importers + tests
updated to the barrel path. Verified: handlers + handlers-stream
suites pass (129 tests).
